requirements.txt
mcp-sdk>=0.1.0               # MCP Python SDK
pydantic>=1.10
httpx>=0.24
pytest>=7.0
pyyaml>=5.4.1                # For loading YAML config